requirements.txt
numpy==2.2.5
tqdm==4.67.1
pandas==2.2.3
matplotlib==3.10.1
opencv-python==4.11.0.86

albumentations==2.0.6
scikit-image==0.25.2
scikit-learn==1.6.1

torch==2.4.1
torch-directml==0.2.5.dev240914
torchmetrics==1.7.1
torchvision==0.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
scipy==1.15.2
ultralytics==8.3.164
seaborn==0.13.2
openvino-dev
onnx
